collectors/news_scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
redis==5.0.1
python-dateutil==2.8.2
langdetect>=1.0.9
lxml==5.1.0